+++ index-v-core.js
@@ -1,0 +1,8 @@
+module.exports = function (Money, repay, ru, about) {
+  return {
+    Money: Money,
+    repay: repay,
+    ru: ru,
+    about: about
+  };
+};