+++ index.js
@@ -1,3 +1,0 @@
-var fiCalc;
-
-module.exports = fiCalc;