+++ libs/repay/repay-i-v-anima-caculate.js
@@ -1,0 +1,1 @@
+module.exports = require('./repay-i-v-core')(require('../money-v-anima-caculate'), require('./repay-util-v-anima-caculate'));