+++ libs/repay/repay-i-v-bignumber.js
@@ -1,0 +1,1 @@
+module.exports = require('./repay-i-v-core')(require('../money-v-bignumber'), require('./repay-util-v-bignumber'));