+++ libs/repay/repay-mpi-v-anima-caculate.js
@@ -1,0 +1,1 @@
+module.exports = require('./repay-mpi-v-core')(require('../money-v-anima-caculate'), require('./repay-util-v-anima-caculate'));