+++ libs/repay/repay-util-v-bignumber.js
@@ -1,0 +1,1 @@
+module.exports = require('./repay-util-v-core')(require('../money-v-bignumber'), require('../util-v-bignumber'));