+++ HISTORY.md
@@ -2,6 +2,16 @@
 
 ---
 
+## 1.0.0
+
+`NEW` 支持还款计划(等额本息, 等额本金, 按月付息到期还本).
+
+`NEW` 支持字符串格式化(千分, 前后缀).
+
+`NEW` 支持取整方式设置(四舍五入, 去尾, 进一).
+
+底层数学库默认依赖 [anima-caculate](http://spmjs.io/package/anima-caculate), 也可选择依赖 [decimal.js](http://spmjs.io/package/decimal.js) 或 [bignumber.js](http://spmjs.io/package/bignumber.js) 或 [big.js](http://spmjs.io/package/big.js).
+
 ## 0.0.0
 
-`new` It is the first version of fi-calc.
+新建仓库, 空内容.